# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.AdminLoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
...
  handlers.AuthResponse:
    properties:
      employee:
        $ref: '#/definitions/models.Employee'
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid credentials
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      nrc:
        example: 123456/78/9
        type: string
      password:
        example: password123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    type: object
...
  handlers.RegisterRequest:
    properties:
      department:
        example: IT
        type: string
      email:
        example: john@example.com
        type: string
      firstname:
        example: John
        type: string
      lastname:
        example: Doe
        type: string
      nrc:
        example: 123456/78/9
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        example: employee
    required:
    - email
    - firstname
    - lastname
    - nrc
    - password
    type: object
...
  models.AuditAction:
    enum:
    - CREATE
    - APPROVE
    - REJECT
    - CANCEL
    - UPDATE
    - DELETE
    type: string
    x-enum-varnames:
    - AuditActionCreate
    - AuditActionApprove
    - AuditActionReject
    - AuditActionCancel
    - AuditActionUpdate
    - AuditActionDelete
  models.AuditEntityType:
    enum:
    - employee
    - identity
    - employment
    - position
    - document
    - compliance
    - onboarding
    - offboarding
    - lifecycle
    - leave
    - leave_type
    type: string
    x-enum-varnames:
    - AuditEntityEmployee
    - AuditEntityIdentity
    - AuditEntityEmployment
    - AuditEntityPosition
    - AuditEntityDocument
    - AuditEntityCompliance
    - AuditEntityOnboarding
    - AuditEntityOffboarding
    - AuditEntityLifecycle
    - AuditEntityLeave
    - AuditEntityLeaveType
  models.AuditLog:
    properties:
      action:
        $ref: '#/definitions/models.AuditAction'
      changes:
        description: JSON representation of what changed
        type: string
      comment:
        type: string
      created_at:
        type: string
      entity_id:
        type: integer
      entity_type:
        $ref: '#/definitions/models.AuditEntityType'
      id:
        type: integer
      ip_address:
        type: string
      new_values:
        description: JSON representation of new values
        type: string
      old_values:
        description: JSON representation of old values
        type: string
      performed_by:
        type: integer
      performer:
        $ref: '#/definitions/models.Employee'
      request_method:
        type: string
      request_path:
        type: string
      user_agent:
        type: string
    type: object
  models.ComplianceRecord:
    properties:
      created_at:
        type: string
      document:
        $ref: '#/definitions/models.Document'
      document_id:
        type: integer
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      expiry_date:
        type: string
      id:
        type: integer
      issue_date:
        type: string
      last_verified_date:
        type: string
      non_compliance_reason:
        type: string
      notes:
        type: string
      requirement:
        $ref: '#/definitions/models.ComplianceRequirement'
      requirement_id:
        type: integer
      status:
        $ref: '#/definitions/models.ComplianceStatus'
      updated_at:
        type: string
      verified_by:
        type: integer
      verifier:
        $ref: '#/definitions/models.Employee'
    type: object
  models.ComplianceRequirement:
    properties:
      category:
        type: string
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_mandatory:
        type: boolean
      name:
        type: string
      records:
        items:
          $ref: '#/definitions/models.ComplianceRecord'
        type: array
      reminder_days:
        description: days before expiry to send reminder
        type: integer
      updated_at:
        type: string
      validity_period:
        description: in days
        type: integer
    type: object
  models.ComplianceStatus:
    enum:
    - compliant
    - non_compliant
    - pending
    - expired
    type: string
    x-enum-varnames:
    - ComplianceStatusCompliant
    - ComplianceStatusNonCompliant
    - ComplianceStatusPending
    - ComplianceStatusExpired
  models.Document:
    properties:
      created_at:
        type: string
      description:
        type: string
      document_type:
        $ref: '#/definitions/models.DocumentType'
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      expiry_date:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      id:
        type: integer
      is_confidential:
        type: boolean
      issue_date:
        type: string
      mime_type:
        type: string
      status:
        $ref: '#/definitions/models.DocumentStatus'
      tags:
        type: string
      title:
        type: string
      updated_at:
        type: string
      uploaded_by:
        type: integer
      uploader:
        $ref: '#/definitions/models.Employee'
      verified_at:
        type: string
      verified_by:
        type: integer
      verifier:
        $ref: '#/definitions/models.Employee'
    type: object
  models.DocumentStatus:
    enum:
    - active
    - expired
    - pending
    - archived
    type: string
    x-enum-varnames:
    - DocumentStatusActive
    - DocumentStatusExpired
    - DocumentStatusPending
    - DocumentStatusArchived
  models.DocumentType:
    enum:
    - id
    - contract
    - resume
    - certificate
    - license
    - performance
    - disciplinary
    - compliance
    - other
    type: string
    x-enum-varnames:
    - DocumentTypeID
    - DocumentTypeContract
    - DocumentTypeResume
    - DocumentTypeCertificate
    - DocumentTypeLicense
    - DocumentTypePerformance
    - DocumentTypeDisciplinary
    - DocumentTypeCompliance
    - DocumentTypeOther
  models.Employee:
    properties:
      audit_logs:
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      compliance_records:
        items:
          $ref: '#/definitions/models.ComplianceRecord'
        type: array
      created_at:
        type: string
      department:
        type: string
      documents:
        items:
          $ref: '#/definitions/models.Document'
        type: array
      email:
        type: string
      employment:
        $ref: '#/definitions/models.EmploymentDetails'
      employment_history:
        items:
          $ref: '#/definitions/models.EmploymentHistory'
        type: array
      firstname:
        type: string
      id:
        type: integer
      identity:
        $ref: '#/definitions/models.IdentityInformation'
      lastname:
        type: string
      leaves:
        description: Relationships
        items:
          $ref: '#/definitions/models.Leave'
        type: array
      lifecycle_events:
        items:
          $ref: '#/definitions/models.WorkLifecycleEvent'
        type: array
      nrc:
        type: string
      offboarding_process:
        $ref: '#/definitions/models.OffboardingProcess'
      onboarding_process:
        $ref: '#/definitions/models.OnboardingProcess'
      position:
        $ref: '#/definitions/models.Position'
      position_assignments:
        items:
          $ref: '#/definitions/models.PositionAssignment'
        type: array
      position_id:
        type: integer
      role:
        $ref: '#/definitions/models.Role'
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.EmploymentDetails:
    properties:
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      employee_number:
        type: string
      employment_status:
        $ref: '#/definitions/models.EmploymentStatus'
      employment_type:
        $ref: '#/definitions/models.EmploymentType'
      end_date:
        type: string
      hire_date:
        type: string
      id:
        type: integer
      manager:
        $ref: '#/definitions/models.Employee'
      manager_id:
        type: integer
      notice_period:
        description: in days
        type: integer
      probation_end_date:
        type: string
      probation_status:
        type: string
      start_date:
        type: string
      termination_date:
        type: string
      termination_reason:
        type: string
      updated_at:
        type: string
      work_location:
        type: string
      work_schedule:
        type: string
    type: object
  models.EmploymentHistory:
    properties:
      change_date:
        type: string
      change_reason:
        type: string
      changed_by:
        type: integer
      changer:
        $ref: '#/definitions/models.Employee'
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      id:
        type: integer
      new_department:
        type: string
      new_position:
        type: string
      new_status:
        $ref: '#/definitions/models.EmploymentStatus'
      notes:
        type: string
      previous_department:
        type: string
      previous_position:
        type: string
      previous_status:
        $ref: '#/definitions/models.EmploymentStatus'
      updated_at:
        type: string
    type: object
  models.EmploymentStatus:
    enum:
    - active
    - on_leave
    - suspended
    - terminated
    - resigned
    type: string
    x-enum-varnames:
    - EmploymentStatusActive
    - EmploymentStatusOnLeave
    - EmploymentStatusSuspended
    - EmploymentStatusTerminated
    - EmploymentStatusResigned
  models.EmploymentType:
    enum:
    - full_time
    - part_time
    - contract
    - internship
    - consultant
    type: string
    x-enum-varnames:
    - EmploymentTypeFullTime
    - EmploymentTypePartTime
    - EmploymentTypeContract
    - EmploymentTypeInternship
    - EmploymentTypeConsultant
  models.IdentityInformation:
    properties:
      address:
        type: string
      blood_group:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      emergency_contact:
        type: string
      emergency_phone:
        type: string
      emergency_relation:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      gender:
        type: string
      id:
        type: integer
      marital_status:
        type: string
      mobile_number:
        type: string
      nationality:
        type: string
      phone_number:
        type: string
      postal_code:
        type: string
      state:
        type: string
      updated_at:
        type: string
    type: object
  models.Leave:
    properties:
      approved_at:
        type: string
      approved_by:
        type: integer
      approver:
        $ref: '#/definitions/models.Employee'
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      leave_type:
        $ref: '#/definitions/models.LeaveType'
      leave_type_id:
        type: integer
      reason:
        type: string
      rejection_reason:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/models.LeaveStatus'
      updated_at:
        type: string
    type: object
...
  models.LeaveStatus:
    enum:
    - Pending
    - Approved
    - Rejected
    - Cancelled
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusApproved
    - StatusRejected
    - StatusCancelled
  models.LeaveType:
    properties:
      created_at:
        type: string
      id:
        type: integer
      leaves:
        items:
          $ref: '#/definitions/models.Leave'
        type: array
      max_days:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.LifecycleEventType:
    enum:
    - hired
    - onboarded
    - promoted
    - transferred
    - demoted
    - resigned
    - terminated
    - retired
    - offboarded
    - status_change
    type: string
    x-enum-varnames:
    - LifecycleEventHired
    - LifecycleEventOnboarded
    - LifecycleEventPromoted
    - LifecycleEventTransferred
    - LifecycleEventDemoted
    - LifecycleEventResigned
    - LifecycleEventTerminated
    - LifecycleEventRetired
    - LifecycleEventOffboarded
    - LifecycleEventStatusChange
  models.OffboardingProcess:
    properties:
      actual_end_date:
        type: string
      assigned_to:
        type: integer
      assignee:
        $ref: '#/definitions/models.Employee'
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      expected_end_date:
        type: string
      id:
        type: integer
      initiated_by:
        type: integer
      initiator:
        $ref: '#/definitions/models.Employee'
      notes:
        type: string
      reason:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/models.OnboardingStatus'
      tasks:
        items:
          $ref: '#/definitions/models.OffboardingTask'
        type: array
      updated_at:
        type: string
    type: object
  models.OffboardingTask:
    properties:
      assigned_to:
        type: integer
      assignee:
        $ref: '#/definitions/models.Employee'
      completed_by:
        type: integer
      completed_date:
        type: string
      completer:
        $ref: '#/definitions/models.Employee'
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      is_required:
        type: boolean
      notes:
        type: string
      offboarding_process:
        $ref: '#/definitions/models.OffboardingProcess'
      offboarding_process_id:
        type: integer
      order:
        type: integer
      status:
        $ref: '#/definitions/models.OnboardingTaskStatus'
      task_name:
        type: string
      updated_at:
        type: string
    type: object
  models.OnboardingProcess:
    properties:
      actual_end_date:
        type: string
      assigned_to:
        type: integer
      assignee:
        $ref: '#/definitions/models.Employee'
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      expected_end_date:
        type: string
      id:
        type: integer
      initiated_by:
        type: integer
      initiator:
        $ref: '#/definitions/models.Employee'
      notes:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/models.OnboardingStatus'
      tasks:
        items:
          $ref: '#/definitions/models.OnboardingTask'
        type: array
      updated_at:
        type: string
    type: object
  models.OnboardingStatus:
    enum:
    - pending
    - in_progress
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - OnboardingStatusPending
    - OnboardingStatusInProgress
    - OnboardingStatusCompleted
    - OnboardingStatusCancelled
  models.OnboardingTask:
    properties:
      assigned_to:
        type: integer
      assignee:
        $ref: '#/definitions/models.Employee'
      completed_by:
        type: integer
      completed_date:
        type: string
      completer:
        $ref: '#/definitions/models.Employee'
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      is_required:
        type: boolean
      notes:
        type: string
      onboarding_process:
        $ref: '#/definitions/models.OnboardingProcess'
      onboarding_process_id:
        type: integer
      order:
        type: integer
      status:
        $ref: '#/definitions/models.OnboardingTaskStatus'
      task_name:
        type: string
      updated_at:
        type: string
    type: object
  models.OnboardingTaskStatus:
    enum:
    - pending
    - in_progress
    - completed
    - skipped
    type: string
    x-enum-varnames:
    - OnboardingTaskStatusPending
    - OnboardingTaskStatusInProgress
    - OnboardingTaskStatusCompleted
    - OnboardingTaskStatusSkipped
  models.Position:
    properties:
      code:
        type: string
      created_at:
        type: string
      department:
        type: string
      description:
        type: string
      employees:
        items:
          $ref: '#/definitions/models.Employee'
        type: array
      id:
        type: integer
      is_active:
        type: boolean
      level:
        type: string
      max_salary:
        type: number
      min_salary:
        type: number
      reports_to:
        $ref: '#/definitions/models.Position'
      reports_to_position:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.PositionAssignment:
    properties:
      assigned_by:
        type: integer
      assigner:
        $ref: '#/definitions/models.Employee'
      assignment_notes:
        type: string
      created_at:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      position:
        $ref: '#/definitions/models.Position'
      position_id:
        type: integer
      salary:
        type: number
      start_date:
        type: string
      updated_at:
        type: string
    type: object
  models.Role:
    enum:
    - employee
    - manager
    - admin
    type: string
    x-enum-varnames:
    - RoleEmployee
    - RoleManager
    - RoleAdmin
  models.WorkLifecycleEvent:
    properties:
      approved_at:
        type: string
      approved_by:
        type: integer
      approver:
        $ref: '#/definitions/models.Employee'
      completion_date:
        type: string
      created_at:
        type: string
      description:
        type: string
      effective_date:
        type: string
      employee:
        $ref: '#/definitions/models.Employee'
      employee_id:
        type: integer
      event_date:
        type: string
      event_type:
        $ref: '#/definitions/models.LifecycleEventType'
      id:
        type: integer
      initiated_by:
        type: integer
      initiator:
        $ref: '#/definitions/models.Employee'
      is_completed:
        type: boolean
      new_value:
        type: string
      notes:
        type: string
      previous_value:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8070
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    REST API for Leave Management System - part of HRMS
    Allows employees to apply for leaves, track balances, and managers to approve/reject requests.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: HRMS Leave Management API
  version: "1.0"
paths:
...
  /auth/admin/login:
    post:
      consumes:
      - application/json
      description: Authenticate admin with username and password, returns JWT token
      parameters:
      - description: Admin login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Admin login
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate employee or manager with NRC and password, returns
        JWT token
      parameters:
      - description: Login credentials (use NRC)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Employee/Manager login
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new employee account
      parameters:
      - description: Employee registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: NRC or email already exists
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register new employee
      tags:
      - Authentication
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" (without quotes)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"